requirements.txt
opencv-python==4.7.0.68
numpy==1.19.5
pandas==1.2.4
pytesseract==0.3.10
Pillow==9.3.0
scipy==1.7.3